django_app/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
gunicorn>=21.0
python-decouple>=3.8
whitenoise>=6.6
djangorestframework>=3.14
django-cors-headers>=4.3
boto3>=1.34
django-storages>=1.14
Pillow>=10.0
